+++ R/dplyr-verbs.R
@@ -100,7 +100,7 @@ slice.tbl_ts <- function(.data, ..., .preserve = FALSE) {
   if (length(pos) > 1) {
     abort("`slice()` only accepts one expression.")
   }
-  pos_eval <- summarise(as_tibble(.data), .pos_eval = list(!!pos[[1]]))
+  pos_eval <- summarise(as_tibble(.data), .pos_eval = list(!! pos[[1]]))
   ascending <- all(map_lgl(pos_eval[[".pos_eval"]], row_validate))
   by_row(slice, .data, ordered = ascending, interval = NULL, ...,
     .preserve = .preserve)